Fix name not shown for registered users without microphone nor camera
When the peer is a registered user her name is shown when a stream is
added for that peer. However, if the peer has no microphone nor camera
then no stream is added, and thus the name was not shown. Now the name
is shown too when the connection is established to guarantee that it
will be shown even if no stream is added.

In the case of guest users the name is shown when a stream is added for
that peer, but also when a "nickChanged" message is received. That
message is sent when the peers are connected, so the name of the guest
user was already properly set even if the user has no microphone nor
camera.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>