Periodically send the nick to update the avatar of later joining participants.
The nick name below the avatar is distributed through the DataChannel of the
PeerConnection and only sent once during establishment. For the MCU case,
the sending PeerConnection is created once and then never changed when more
participants join.

For this, we periodically send the nick to all other participants through the
sending PeerConnection.

Signed-off-by: Joachim Bauch <bauch@struktur.de>