Add columns "object_type" and "object_id" to table "talk_rooms"
Note that in Oracle string fields have to be nullable to be able to set
empty strings on them.

As the table "comments" also has columns named "object_type" and
"object_id" they must be namespaced when the last comment is included in
the room information; otherwise the created comment object would have
the object type and id from the room instead of from the comment.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>
Signed-off-by: Joas Schilling <coding@schilljs.com>