Adjust main layout to changes in server
Main layout elements (navigation, content and sidebar) are no longer
children of the "#app" element in the standard layout in the server, so
those changes are reflected in Talk too.

Note, however, that the "#app" element is still kept as an, in practice,
invisible element, as it is used as a holder for some values (like the
token or the signaling settings) used in scripts.

Besides that, although the layout of the server is mainly mimicked in
Talk, there are also some differences needed for its specific UI.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>