Fix z-index of the sidebar trigger
The z-index of the "#app-content" is now set in the server to 1000. Due
to the layout changes the trigger is now the child of a sibling of
"#app-content" instead of a descendant of "#app-content", so its z-index
must be higher than the one for "#app-content" to appear in front of it.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>