Fix input field for new conversations and navigation bar
It is no longer needed to use an absolute position with a fixed width
for the input field for new conversations; the default CSS rules will
cause it to fill the navigation bar width.

As the input field now uses a static position it is taken into account
when calculating the position of the conversation list. Therefore there
is no need to use an extra top padding in the navigation bar.

Finally, the default bottom padding for the navigation bar is 0, so
there is no need to set it to that value.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>