Fix step to rename a conversation in acceptance tests
The edit button of editable text labels is always hidden, except when
the cursor is on the editable text label. Due to this the step to rename
a conversation first clicked on the editable text label and then, given
that it was now visible, on the edit button.

Before clicking on an element the cursor is moved to the middle of that
element, and due to the layout changes, it turns out that the edit
button is now at the middle of the editable text label; when the cursor
was moved the edit button was shown and it ended receiving the click
instead of the editable text label. This caused the editable text label
to enter in edit mode, so the edit button was hidden, and then it could
not be clicked in the next action as expected.

Now only the label element is clicked instead of the containing editable
text label; as the label element is a sibling of the edit button shown
next to it and both are children of the editable text label this causes
the edit button to be shown in the first action and then clicked in the
second action, just like before.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>