Fix colour of app navigation toggle when in a call
The app navigation toggle, which is only shown in narrow screens, is a
dark icon defined in the server. During calls or screensharing the
background of the app content is set to black, so the icon was not
visible. Now the icon turns white in those cases, like the sidebar
trigger or the fullscreen icon.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>