Fix colour of own media icons
The icons used for own media actions (muting, disabling the video or
screensharing) are usually white, except when there is no call, no
screensharing and no local video, in which case they appear on a white
background. In that case they should be black instead, so the default
white images are overridden with their black variant.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>