Setup signaling event handlers only once
Every time that a conversation to request a password was successfully
created a new handler for the signaling event was registered. Thus, if a
password was requested, the conversation ended, and the password was
requested again from the same page then each event had two handlers.

Until now, due to the behaviour of the handlers, even if incorrect it
was not a problem. However, in preparation for changes to the "joinRoom"
handler now a single handler is registered for each signaling event, no
matter how many passwords were requested.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>