Replace call to "connection.joinRoom" with custom implementation
The implementation of "connection.joinRoom" calls
"app.syncAndSetActiveRoom", which performs several actions needed when
the full Talk UI is used, but unneeded when the embedded Talk sidebar is
used.

Moreover, as it is not possible to know when the rooms were synced,
joining the call after joining the room could lead to UI issues due to
the order in which elements were setup (for example, the empty content
message "waiting for permissions" could be shown instead of "waiting for
XXX to join the call"), so this is also a needed change to be able to
add support for video calls.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>