Add basic support for video calls
When a password is requested now the guest automatically joins the call,
and once the sharer joins the call too a video call view appears in the
Talk sidebar.

Although it is not currently shown, the empty content message for guest
users was set, as it is expected to be set by some event handlers.

In a similar way the "#screens" element was also added, but there is no
support yet for screensharing and thus the element is kept always
hidden.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>