Show empty content message when not in a call
The empty content message informs the user when the app is waiting for
the media permissions and then when it is waiting for the sharer to join
the call. Thus it is now shown above the chat view, in the place that
will be occupied by the call container once the call starts.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>