Work around fetching participants before setting the room
The room data currently comes from both the signaling and queries to the
server, and each one of those sources have its own set of events. Due to
this, when entering a room the list of participants can sometimes be
tried to be fetched before the room was set for the list of
participants.

This commit works around that problem by fetching the participant list
only when a URL was set; the real fix (proper synchronization between
the events from both sources) would require a much deeper overhaul of
the architecture.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>