Do not try to sync rooms before requesting the password
In the public page for a conversation the token is always defined, so
the signaling can be initialized as soon as the app starts. However, in
the public share auth page the token is only available once the password
is requested and the room created, so trying to initialize the signaling
when the app starts ends in failed requests to the room endpoint.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>