Add favourite icon to the conversation list
Favourited conversations now show a favourite icon in the top right of
their avatar/conversation type icon in the conversation list.

The structure is based on the one used for the file list by the Tags
plugin of the Files app. However, the "favorite" icon was used instead
of the "starred" icon, as in this case it is just a mark that can not be
interacted with, and the "starred" icon changes its appearance on hover
and focus.

The mark can not be added as a child of the avatar/conversation type
icon, as it would be removed when the avatar is loaded. Therefore it is
a sibling of it placed on its top right corner using absolute
positioning.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>