Add button to share files from Talk UI
The button, which is shown next to the area to write new messages, shows
a file picker and then shares the chosen file with the room.

Note that currently new files can not be uploaded through the file
picker; only those already in the Nextcloud account of the user can be
chosen.

Also, currently only registered users can share files with a room, so
the button is hidden for guests.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>