Make possible to run Docker integration tests on PostgreSQL
Now it is possible to use PostgreSQL as the database when running the
integration tests on Docker by calling "run-docker.sh" with the
"--database pgsql" option.

It is not possible to wait for the PostgreSQL server to be ready using
curl like it was done for MySQL; when sending a HTTP request to the
PostgreSQL port the server sends an empty response, which curl sees as
an error. However, as Bash supports testing a connection with a remote
TCP socket, and this approach is valid for both MySQL and PostgreSQL
servers, the curl call was replaced with this Bash construct.

The default PostgreSQL database image is "postgres:10", the same
currently used when running the tests on Drone; a different one can be
specified with the "--database-image IMAGE_NAME" option.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>