Make possible to run Docker integration tests on MySQL
Now it is possible to use MySQL as the database when running the
integration tests on Docker by calling "run-docker.sh" with the
"--database mysql" option.

This will launch the database server in its own Docker container,
connect it with the container in which Nextcloud is run, and set the
database as the one to be used by Nextcloud.

If "--database XXX" is not specified then SQLite is used; in that case
everything behaves as before (no database server is started).

The default MySQL database image is "mysql:5.7", the same currently used
when running the tests on Drone; a different one can be specified with
the "--database-image IMAGE_NAME" option.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>