Ensure that "object_type" and "object_id" are strings
As "object_type" and "object_id" are nullable the row returned by the
database can contain a null value in those fields. However, those fields
are nullable only due to some Oracle limitations; conceptually they
should be empty strings, so an explicit cast to string is now used when
creating the room object.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>