Fix position of contacts menu shown on mentions
The server does not provide default CSS rules for the contacts menu, as
its position depends on the element on which it is shown.

The contacts menu is now positioned to show the tip of the arrow
horizontally aligned with the center of the avatar, and with the top of
the menu slightly below the bottom border of the mention.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>