Do not render again the participant list on every collection event
The participant list view was rendered on every collection event to
overcome the problem of participant views not being rendered, but that
caused unneeded renderings, for example when the collection was synced
but it had not changed. Now that the id of models in the participant
collection is properly set the view for each participant is rendered
again as expected when the attributes in its model change, so it is no
longer needed to render the participant list on every collection event.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>