Set the attribute model on a successful operation
The endpoints for moderation operations do not fit the default setup for
Backbone models, so the operations are performed directly with AJAX
requests on the view. Thus, after a successful request rendering again
the view has no visible effect, as the model it is based on has not
changed yet. Now the model is updated instead, which will trigger the
render with the new state; that state will be later refreshed with the
server state (although it should be the same) when the model is synced
again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>