Fix rendering of online/offline status
When a participant is offline the ".participant-offline" CSS class is
added to the element of its view. However, when the view is rendered
again the element itself is reused and only its contents are replaced by
the template. Thus, the CSS class must be explicitly removed from the
element when the view is rendered if the user is now online.

As the CSS class is expected (by the CSS rules) to be set on the
participant element instead of on the link inside it the CSS class was
removed from the template. This makes no difference, though, as
"isOffline" was always undefined and thus the class was never set on the
link.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>