Use a document fragment as a buffer for new message DOM elements
Each time a new node is added to the document a reflow and a render is
triggered. This can cause the UI to become unresponsive when a lot of
nodes are added in a row, for example, when loading the messages for the
first time after entering a room.

To mitigate that now the DOM elements for the messages to add are
buffered using a document fragment. Instead of adding the element for
each model directly to the document now they are added to the document
fragment, which does not cause a reflow nor a render, and once all the
elements have been created then they are added at once to the document,
which causes a single reflow and render instead of one for each element.

Despite that, note that when there are thousands of messages in the list
even a single reflow and render can make the UI hang for a few seconds;
using a virtual list is probably the only way to solve this.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>