Get shared with e-mail from the share instead of from the room name
This prepares the code for future changes in which the password could be
requested for other types of shares.

Note that the notification of a password request is just a regular call
notification; thus, the share is not provided in the notification
parameters and it has to be got from the ID of the object associated to
the room (as in "share:password" rooms the object ID is the share
token).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>