Fix room name in CallInfoView for "share:password" rooms
The name of rooms for requesting a password by Talk is generated by the
server each time a room or the list of rooms is fetched, and it should
not be editable by the user. Moreover, the "name" attribute is the raw
name from which the "displayName" attribute is generated, and thus the
later is the one that should be shown for "share:password" rooms.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>