Fix the scroll position of the chat view in the sidebar
When the chat view was in the sidebar, the scroll position of the chat
view was reset to the top every time that the sidebar was closed and
opened again or the chat tab was selected after changing to a different
one.

The reason is that detaching an element and attaching it again, which
happens in both scenarios, resets the scroll position of the element, so
now the scroll position is explicitly saved before closing the sidebar
or unselecting the chat tab and restored again after opening the sidebar
or selecting the chat tab again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>