Make possible to restore the scroll position to the last known position.
In some cases there is no event to listen to for a change that affects
the chat view, so it is not possible to explicitly save the scroll
position when that change happens. Now the virtual list keeps track of
the last known scroll position, and the chat view allows restoring to an
arbitrary scroll position besides the saved one. This makes possible,
for example, to restore to the last scroll position before the chat view
was hidden (as long as it is possible to detect that the chat view is
shown again, even if it was not possible to detect that it was hidden).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>