Show Talk sidebar in Files app
This commit introduces a DetailTabView plugin to show a chat view in the
sidebar of the Files app. The tab makes possible to chat in a Talk room
associated to the current file; due to this, the tab is visible only on
files that can be associated to a room, that is, files shared with the
current user or by the current user to another user (as a user, group,
circle...).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>