Replace "videoWasEnabledAtLeastOnce" with "videoNotFound"
"videoWasEnabledAtLeastOnce" was used to allow or prevent the handling
of clicks on the video button. However, it was enabled in the handler
for the "localstreams" event, so it was being enabled even if there was
no video (although it was disabled as needed later during the handling
of the "localMediaStarted" event). In any case, its purpose can be
covered by "videoNotFound", so it was removed in favour of the last.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>