Unify handling of "localstream" with "localMediaStarted"
SimpleWebRTC emits the "localstream" event after a successful call to
"start/startLocalMedia". In turn, this is only called from
"startLocalVideo", which emits "localMediaStarted" if "startLocalMedia"
succeeded; "localMediaStarted" is handled by calling "startLocalMedia"
in the app, so now the setup done when handling "localstream" is merged
with the setup done when handling "localMediaStarted".

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>