Move MediaControlsView to precompiled Handlebars templates
In Nextcloud 15 the default Content Security Policy disallows unsafe
eval expressions, so Handlebars templates can no longer be compiled at
runtime.

For the time being that default Content Security Policy was lifted for
Talk so "Handlebars.compile" could still be used. However, this only
applies to Talk itself; when using Talk components in other apps they
must abide to the Content Security Policy of those apps. As
MediaControlsView is going to be used in the Files app it has been moved
to precompiled Handlebars templates (which are still compatible with the
regular Talk UI).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>