Fix room polling not stopped when leaving the current room
When setting a room in the internal signaling a poller is started that
synchronizes the room data periodically; that poller should be stopped
when the room is left.

Note, however, that the poller should not be stopped when leaving the
current room if it was initialized for a room collection instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>