Base message on number of users and additional message on type of room
Before, the message explicitly mentioned the other user and her avatar
was used as the icon only for one to one rooms. Now that will be done
for any type of room if there is only another registered user and no
guests in that room.

In a similar way, the additional message hinting how to invite other
users now is always shown in group and public rooms, no matter how many
users are currently in the room; this is the same as before for public
rooms, although it was changed for consistency for group rooms, as in
that case the message was shown only when there were no other
participants in the room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>