Set "WebRTC is not supported" message on every call
Before, the empty content message informing the user that WebRTC was not
supported was set only when WebRTC was initialized. Due to this the
message could be overwritten when changing to a different room, and when
starting/joining a call in that other room the "Waiting for others"
message was shown instead. Now the "WebRTC is not supported" messages is
set on every call to prevent that.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>