Set the message explicitly instead of restoring the previous one
Restoring the message once the permissions were granted or not could
lead to the wrong message being shown if a different message was set in
the meantime; this would typically not happen in the general UI, but it
was a problem that had to be worked around in the "Request password" UI
by not using the main "syncAndSetActiveRoom" method.

As the message to be restored is deterministic now it is explicitly set
instead; it will always be the "Waiting for others to join the call"
message, except when WebRTC is not supported which will not be modified.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>