Move changes to the session history from RoomListView to Connection
The session history is modified by Connection when leaving the current
room for any reason (either because the user explicitly left the room or
because she no longer has access to it), so there is no need to modify
the session history from RoomListView.

For consistency, the session history is now modified only from
Connection when joining a room too, instead of modifying it from both
the RoomListView and Connection.

In any case, modifying the session history should probably done from App
instead as a response to an event, but for the time being this is good
enough.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>